requirements.txt
mysql-connector-python == 8.0.23
pymongo == 3.11.3
neo4j == 4.2.1
flask == 1.1.2
redis == 3.5.3
